Separate mappings into individual files for each pgn
