+++ mmsi-context.js
@@ -1,0 +1,5 @@
+
+module.exports = function(n2k) {
+  return 'vessels.urn:mrn:imo:mmsi:' + n2k.fields['User ID'];
+}
+
