+++ package.json
@@ -23,7 +23,8 @@
     "debug": "^2.1.1",
     "nomnom": "~1.6.2",
     "signalk-schema": "SignalK/specification",
-    "through": ">=2.2.7 <3"
+    "through": ">=2.2.7 <3",
+    "glob": "^7.7.1"
   },
   "devDependencies": {
     "chai": "1.9",
