+++ pgns/127245.js
@@ -1,11 +1,10 @@
 
-exports.mappings = {
-  //Rudder
-  '127245': [{
+module.exports = [
+  {
     source: 'Position',
     node: 'steering.rudderAngle',
     filter: function(n2k) {
       return typeof n2k.fields['Position'] != 'undefined'
     }
-  }]
-}
+  }
+]
