+++ pgns/127251.js
@@ -1,0 +1,7 @@
+
+module.exports = [
+  {
+    source: 'Rate',
+    node: 'navigation.rateOfTurn'
+  }
+]
