+++ pgns/127257.js
@@ -1,0 +1,13 @@
+
+module.exports = [
+  {
+    value: function(n2k) {
+      return {
+        yaw: Number(n2k.fields.Yaw),
+        pitch: Number(n2k.fields.Pitch),
+        roll: Number(n2k.fields.Roll)
+      }
+    },
+    node: 'navigation.attitude'
+  }
+]
