+++ pgns/127258.js
@@ -1,0 +1,7 @@
+
+module.exports = [
+  {
+    source: 'Variation',
+    node: 'navigation.magneticVariation'
+  }
+]
