+++ pgns/127489.js
@@ -1,27 +1,6 @@
 
-exports.mappings = {
-  // Engine speed data
-  '127488': [{
-    node: 'propulsion.port.revolutions',
-    filter: function(n2k) {
-      return n2k.fields['Engine Instance'] === 'Single Engine or Dual Engine Port';
-    },
-    value: function(n2k) {
-      var rpm = Number(n2k.fields['Engine Speed'])
-      return rpm / 60.0;
-    }
-  }, {
-    node: 'propulsion.starboard.revolutions',
-    filter: function(n2k) {
-      return n2k.fields['Engine Instance'] === 'Dual Engine Starboard';
-    },
-    value: function(n2k) {
-      var rpm = Number(n2k.fields['Engine Speed'])
-      return rpm / 60.0;
-    }
-  }],
-  // Engine operating parameters
-  '127489': [{
+module.exports = [
+  {
     source: 'Temperature',
     node: 'propulsion.port.temperature',
     filter: function(n2k) {
@@ -87,5 +66,5 @@ exports.mappings = {
     filter: function(n2k) {
       return n2k.fields['Engine Instance'] === 'Dual Engine Starboard';
     }
-  }]
-}
+  }
+]
