+++ pgns/128259.js
@@ -1,0 +1,16 @@
+
+module.exports = [
+  {
+    source: 'Speed Water Referenced',
+    node: 'navigation.speedThroughWater',
+    filter: function(n2k) {
+      return n2k.fields['Speed Water Referenced']
+    }
+  }, {
+    source: 'Speed Ground Referenced',
+    node: 'navigation.speedOverGround',
+    filter: function(n2k) {
+      return n2k.fields['Speed Ground Referenced']
+    }
+  }
+]
