+++ pgns/128275.js
@@ -1,0 +1,10 @@
+
+module.exports =  [
+  {
+    source: 'Trip Log',
+    node: 'navigation.logTrip'
+  }, {
+    source: 'Log',
+    node: 'navigation.log'
+  }
+]
