+++ pgns/129025.js
@@ -1,0 +1,12 @@
+
+module.exports = [
+  {
+    value: function(n2k) {
+      return {
+        longitude: Number(n2k.fields.Longitude),
+        latitude: Number(n2k.fields.Latitude)
+      }
+    },
+    node: 'navigation.position'
+  }
+]
