+++ pgns/129291.js
@@ -1,0 +1,20 @@
+
+module.exports = [
+  {
+    node: 'environment.current',
+    value: function(n2k) {
+      if (n2k.fields['Set Reference'] === 'True') {
+        return {
+          setTrue: Number(n2k.fields.Set),
+          drift: Number(n2k.fields['Drift'])
+        };
+      } else if (n2k.fields['Set Reference'] === 'Magnetic') {
+        //speculative, I don't have a real world sample showing 'Magnetic'
+        return {
+          setTrue: Number(n2k.fields.Set),
+          drift: Number(n2k.fields['Drift'])
+        };
+      }
+    }
+  }
+]
